KFLUXINFRA-2497 - Pin oauth2-proxy submodule to stable tag (v7.12.0)
The git submodule pointing to the upstream oauth2-proxy repository was
previously following the 'master' branch, which caused MintMaker/Renovate
to create a new PR each time that branch was updated.

To align with Konflux, where the build pipeline uses stable versions, this
change pins the submodule in the .gitmodules file to the latest stable tag
(v7.12.0), following Renovate documentation [1]. This minimizes the number
of open PRs.

As this change disrupts the local git submodule workflow, instructions on
how to mitigate the issue have been added to the README file.

[1] https://docs.renovatebot.com/modules/manager/git-submodules/#updating-to-specific-tag-values